refactor: 查询单词
